+++ src/backoffice/SlidesScreen/ItemList.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export const ItemList = () => {
+    return (
+        <div>
+            <h1>Item list</h1>
+            <p> SEXOOO </p>
+        </div>
+    )
+}
